refactor: huya model refactor complete
