--- main.ts
@@ -1,5 +1,0 @@
-import { app, BrowserWindow } from "electron";
-import getHuyaSteam from "./engine/huya/message";
-import * as path from "path";
-
-getHuyaSteam();
