+++ src/index.ts
@@ -1,0 +1,3 @@
+import { getHuyaSteam } from "./engine/huya/message";
+
+getHuyaSteam();
