+++ src/type/getHuya.ts
@@ -1,0 +1,4 @@
+export interface HuyaStreamInfo {
+  streamUrl: string;
+  streamName: string;
+}
