+++ templates/info.json
@@ -4,6 +4,6 @@
     "password": "xx"
   },
   "streamerInfo": {
-    "1号主播": "https://www.huya.com/626813"
+    "奶音Lolita小野猫第一天直播": "https://www.huya.com/21467113"
   }
 }
