--- types/message.d.ts
@@ -1,1 +1,0 @@
-type a = number;
